server/requirements.txt
fastapi~=0.115.0
openai~=1.50.2
python-dotenv~=1.0.1
numpy~=2.0.2
torch~=2.4.1
uvicorn~=0.31.0
python-multipart~=0.0.12
pydantic~=2.8.2
librosa~=0.10.2.post1
requests~=2.32.3